Cast to uint32_t a bunch to remove compile warnings

+++ src/trace/cell-tracing.c
@@ -83,8 +83,10 @@ void tor_trace_channel_tls_write_packed_cell(connection_t *conn,
   int64_t diff = ( ((uint64_t)ts.tv_sec) *1000000000 + ts.tv_nsec ) -
     ( ((uint64_t)c_info->ts.tv_sec)*1000000000 + c_info->ts.tv_nsec );
   log_info(LD_OR, "[cell-tracing] %" PRIu32 ".%" PRIu32 " %" PRIu32 ".%" PRIu32
-      " %" PRIi64 " id=%" PRIu32 " waiting in outbuf", c_info->ts.tv_sec,
-      c_info->ts.tv_nsec, ts.tv_sec, ts.tv_nsec, diff, c_info->id);
+      " %" PRIi64 " id=%" PRIu32 " waiting in outbuf",
+      (uint32_t)c_info->ts.tv_sec,
+      (uint32_t)c_info->ts.tv_nsec, (uint32_t)ts.tv_sec, (uint32_t)ts.tv_nsec,
+      diff, c_info->id);
 }
 
 void tor_trace_connection_cell_inbuf(cell_t *cell)
@@ -96,7 +98,7 @@ void tor_trace_connection_cell_inbuf(cell_t *cell)
     cell->id = cell_next_id++;
     cell_counter -= trace_every_n;
     log_info(LD_OR, "[cell-tracing] %" PRIu32 ".%" PRIu32 " id=%" PRIu32,
-        cell->ts.tv_sec, cell->ts.tv_nsec, cell->id);
+        (uint32_t)cell->ts.tv_sec, (uint32_t)cell->ts.tv_nsec, cell->id);
   } else {
     cell->id = 0;
     cell->ts.tv_sec = cell->ts.tv_nsec = 0;
@@ -130,8 +132,9 @@ void tor_trace_connection_write_to_buf_flushed(connection_t *conn, int amount)
         ( ((uint64_t)c_info->ts.tv_sec)*1000000000 + c_info->ts.tv_nsec );
       log_info(LD_OR, "[cell-tracing] %" PRIu32 ".%" PRIu32 " %" PRIu32
           ".%" PRIu32 " %" PRIi64 " id=%" PRIu32 " written to kernel",
-          c_info->ts.tv_sec, c_info->ts.tv_nsec, ts.tv_sec, ts.tv_nsec, diff,
-          c_info->id);
+          (uint32_t)c_info->ts.tv_sec,
+          (uint32_t)c_info->ts.tv_nsec, (uint32_t)ts.tv_sec,
+          (uint32_t)ts.tv_nsec, diff, c_info->id);
       tor_free(c_info);
       MAP_DEL_CURRENT(k);
     }
